+++ InventoryTools/Logic/ItemRenderers/ItemDungeonChestSourceRenderer.cs
@@ -35,11 +35,12 @@ public class ItemDungeonChestSourceRenderer : ItemInfoRenderer<ItemDungeonChestS
     public override Action<ItemSource> DrawTooltip => source =>
     {
         var asSource = AsSource(source);
+
         ImGui.Text("Dungeon: " + asSource.ContentFinderCondition.FormattedName);
         using (ImRaii.PushIndent())
         {
             ImGui.Text(
-                $"Chest {asSource.DungeonChest.ChestNo + 1} ({asSource.DungeonChest.Position.X} / {asSource.DungeonChest.Position.Y})");
+                $"Chest {asSource.DungeonChest.ChestNo + 1})");
         }
     };
 
@@ -62,7 +63,7 @@ public class ItemDungeonChestSourceRenderer : ItemInfoRenderer<ItemDungeonChestS
                 foreach (var chest in dungeon.OrderBy(c => c.DungeonChest.ChestNo))
                 {
                     ImGui.Text(
-                        $"Chest {chest.DungeonChest.ChestNo + 1} ({chest.DungeonChest.Position.X} / {chest.DungeonChest.Position.Y})");
+                        $"Chest {chest.DungeonChest.ChestNo + 1}");
                 }
             }
         }
@@ -82,6 +83,6 @@ public class ItemDungeonChestSourceRenderer : ItemInfoRenderer<ItemDungeonChestS
     public override Func<ItemSource, string> GetDescription => source =>
     {
         var asSource = AsSource(source);
-        return $"{asSource.ContentFinderCondition.FormattedName} ({asSource.DungeonChest.Position.X } / {asSource.DungeonChest.Position.Y})";
+        return $"{asSource.ContentFinderCondition.FormattedName}";
     };
 }