+++ InventoryTools/Logic/ItemRenderers/ItemSupplementSourceRenderer.cs
@@ -286,6 +286,18 @@ public class ItemHydatosSourceRenderer : ItemSupplementSourceRenderer<ItemHydato
     public override string HelpText => "Can the item be obtained from a loot item in Eureka Hydatos?";
 }
 
+public class ItemPilgrimsTraverseSourceRenderer : ItemSupplementSourceRenderer<ItemPilgrimsTraverseSource>
+{
+    public override IReadOnlyList<ItemInfoRenderCategory>? Categories { get; } =
+        [ItemInfoRenderCategory.FieldOperation, ItemInfoRenderCategory.DeepDungeon];
+    public ItemPilgrimsTraverseSourceRenderer(ItemSheet itemSheet, MapSheet mapSheet, ITextureProvider textureProvider, IDalamudPluginInterface pluginInterface) : base(itemSheet, mapSheet, textureProvider, pluginInterface,  ItemInfoType.PilgrimsTraverse, Icons.FieldOpsIcon)
+    {
+    }
+
+    public override string SingularName => "Pilgrim's Traverse";
+    public override string HelpText => "Can the item be obtained from a loot item in Pilgrim's Traverse?";
+}
+
 public class ItemHydatosUseRenderer : ItemSupplementUseRenderer<ItemHydatosSource>
 {
     public override IReadOnlyList<ItemInfoRenderCategory>? Categories { get; } =